# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for ItsPlanned - A Collaborative Event Planning Application. This API provides endpoints for managing events, tasks, user profiles, and integrations with external services like Google Calendar.",
        "title": "ItsPlanned API",
        "contact": {
            "name": "ItsPlanned Support",
            "url": "https://github.com/vl4ddos/itsplanned",
            "email": "support@itsplanned.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/auth/google": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the URL for Google OAuth authorization",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "Get Google OAuth URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Custom redirect URI",
                        "name": "redirect_uri",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OAuth URL generated successfully",
                        "schema": {
                            "$ref": "#/definitions/api.GoogleOAuthURLResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/auth/google/callback": {
            "get": {
                "description": "Handle the callback from Google OAuth and exchange code for tokens",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "Google OAuth callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization code from Google",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Custom redirect URI",
                        "name": "redirect_uri",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Deeplink URI to redirect to after OAuth",
                        "name": "app_redirect",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tokens received successfully",
                        "schema": {
                            "$ref": "#/definitions/api.GoogleOAuthCallbackResponse"
                        }
                    },
                    "302": {
                        "description": "Redirect to app with tokens",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Authorization code not provided",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get access token",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/web-to-app": {
            "get": {
                "description": "Redirects from web OAuth flow to mobile app via deeplink",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "OAuth Web to App Redirect",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization code from Google",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "State parameter for security",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to app deeplink",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/events/invite": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate a unique invite link for an event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invitations"
                ],
                "summary": "Generate event invite link",
                "parameters": [
                    {
                        "description": "Event ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.GenerateInviteLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invite link generated successfully",
                        "schema": {
                            "$ref": "#/definitions/api.GenerateInviteLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/events/join/{invite_code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Join an event using a unique invite code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invitations"
                ],
                "summary": "Join event using invite link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invite Code from path",
                        "name": "invite_code",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Invite Code from query parameter",
                        "name": "code",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully joined event",
                        "schema": {
                            "$ref": "#/definitions/api.JoinEventResponse"
                        }
                    },
                    "400": {
                        "description": "Already a participant",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Invalid invite link",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/events/redirect/{invite_code}": {
            "get": {
                "description": "Redirects from web browser to iOS app deeplink for event joining",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "invitations"
                ],
                "summary": "Redirect to iOS app deeplink",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invite Code",
                        "name": "invite_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to iOS app",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Invalid invite link",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/events/{id}/leave": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allow a user to leave an event they are participating in",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invitations"
                ],
                "summary": "Leave an event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully left event",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid event ID",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Cannot leave - you are the organizer",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not a participant or event not found",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid input"
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "user": {
                    "$ref": "#/definitions/api.UserResponse"
                }
            }
        },
...
        "api.GenerateInviteLinkRequest": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "api.GenerateInviteLinkResponse": {
            "type": "object",
            "properties": {
                "invite_link": {
                    "type": "string",
                    "example": "http://localhost:8080/events/join/abc123"
                }
            }
        },
        "api.GoogleOAuthCallbackResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "ya29.a0AfB_byC..."
                },
                "expiry": {
                    "type": "string",
                    "example": "2024-03-16T15:04:05Z"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "1//04dK..."
                }
            }
        },
        "api.GoogleOAuthURLResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string",
                    "example": "https://accounts.google.com/o/oauth2/auth?..."
                }
            }
        },
...
        "api.JoinEventResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Successfully joined event"
                }
            }
        },
...
        "api.UserResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "bio": {
                    "type": "string",
                    "example": "Software developer and tech enthusiast"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-03-16T12:00:00Z"
                },
                "display_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-03-16T12:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}